Check Point - Sleep
Minor corrections, and additional expansion on UTF and the mess it makes and how this travesty still needs a fix.  Updated a few things... and marked a checkpoint so I can find it easily later as to where I left off with basically another pass of proofreads to that point.